refactor: remove unnecessary newlines, move pool utils to utilities.ts
