+++ src/__playground__/index.ts
@@ -1,5 +1,10 @@
+import { Messenger } from 'nanolith';
 import { worker } from './worker.js';
+
+const foo = new Messenger('foo');
 
 const service = await worker.launchService();
 
-await service.call({ name: 'exit1' });
+await service.sendMessenger(foo);
+
+console.log('resolve');
