+++ src/constants/pool.ts
@@ -44,8 +44,8 @@ export enum ConcurrencyOption {
 export const concurrencyOptionMultipliers = {
     [ConcurrencyOption.Quarter]: 0.25,
     [ConcurrencyOption.Half]: 0.5,
-    [ConcurrencyOption.Default]: 2,
     [ConcurrencyOption.x1]: 1,
+    [ConcurrencyOption.Default]: 2,
     [ConcurrencyOption.x2]: 2,
     [ConcurrencyOption.x4]: 4,
     [ConcurrencyOption.x6]: 6,
